perf: 서열표 필터 전환 호출 제거
